# core/vm: speed up push and interpreter loop