--- a/core/vm/instructions_test.go
+++ b/core/vm/instructions_test.go
@@ -928,7 +928,7 @@ func TestOpMCopy(t *testing.T) {
 	}
 }
 
-// TestPush sanity-checks how code with immediates are handled when the code size is 
+// TestPush sanity-checks how code with immediates are handled when the code size is
 // smaller than the size of the immediate.
 func TestPush(t *testing.T) {
 	code := common.FromHex("0011223344556677889900aabbccddeeff0102030405060708090a0b0c0d0e0ff1e1d1c1b1a19181716151413121")