@holiman 2024-10-23
Btw if we want this, we can also remove `tracing.GasChangeIgnored`. cc @s1na ?